requirements.txt
requests
pyyaml
python-dateutil
rapidfuzz
feedparser
beautifulsoup4
sqlite-utils
schedule
